Move main function to new file
